backend/requirements.txt
# Django Core
Django==5.0.3
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1

# Database
psycopg2-binary==2.9.9

# CORS
django-cors-headers==4.3.1

# Environment variables
python-dotenv==1.0.1

# Payment Gateway
razorpay==1.4.2

# Server (production)
gunicorn==21.2.0

# Image/File handling
Pillow==10.2.0
